% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anciennes_nouvelles_regions}
\alias{anciennes_nouvelles_regions}
\title{anciennes_nouvelles_regions}
\format{
Un data frame de 101 lignes et 6 colonnes :
\describe{
\item{num_dep}{numero du departement.}
\item{dep_name}{nom du departement.}
\item{nouvelles_regions}{nom de la nouvelle region.}
\item{anciennes_regions}{nom de l'ancienne region.}
\item{code_insee_nouvelle_region}{code insee de la nouvelle region.}
\item{code_insee_departement}{code insee du departement.}
}
}
\usage{
anciennes_nouvelles_regions
}
\description{
Data frame contenant la correspondance entre anciennes et nouvelles regions
}
\keyword{datasets}
