% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{enjeu}
\alias{enjeu}
\title{Calculer l'enjeu de conservation d'une espece}
\usage{
enjeu(lr_nat, lr_reg = NA, reg_nat)
}
\arguments{
\item{lr_nat}{Statut dans la liste rouge nationale (\code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"DD"}, \code{"NA"}, \code{"NE"}).
Ce parametre est obligatoire.}

\item{lr_reg}{Statut dans la liste rouge regionale (\code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"DD"}, \code{"NA"}, \code{"NE"}).
Par defaut, il est defini a \code{NA}.}

\item{reg_nat}{ratio du nombre d'observations regionales avec le nombre d'observation nationales  (\code{numeric}).
Ce parametre est obligatoire.}
}
\value{
Un score numerique representant l'enjeu de conservation.
Le score est calcule comme suit :
\deqn{\text{enjeu} = \text{vulnerabilite} \times \text{reg\_nat} \times 10 \times 2}
}
\description{
Cette fonction evalue l'enjeu de conservation d'une espece en combinant son niveau de vulnerabilite,
son importance regionale par rapport au niveau national, et un facteur multiplicatif fixe.
}
\details{
La fonction utilise le score de vulnerabilite obtenu via la fonction \code{\link{vulnerabilite}} et applique un
coefficient d'importance regionale (\code{reg_nat}) pour ponderer l'enjeu de conservation. Le score final est amplifie
par un facteur fixe de 20 (\code{10 * 2}).
}
\examples{
\dontrun{
# Exemple avec un statut national "VU" et un ratio regional-national de 0.8
enjeu_score <- enjeu("VU", "EN", reg_nat = 0.8)
print(enjeu_score)

# Exemple sans statut regional
enjeu_score2 <- enjeu("VU", reg_nat = 0.8)
print(enjeu_score2)  # Retourne un score base uniquement sur le statut national
}

}
\seealso{
\code{\link{vulnerabilite}}
}
