% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{match_inpn}
\alias{match_inpn}
\title{Identifier l'ID INPN et le nom vernaculaire d'une espece}
\usage{
match_inpn(espece)
}
\arguments{
\item{espece}{Une chaine de caracteres correspondant au nom scientifique de l'espece.
Les espaces en trop seront automatiquement supprimes.}
}
\value{
Une liste contenant :
\itemize{
\item \code{id} : L'identifiant numerique correspondant l'identifiant TaxRef de l'espece.
\item \code{nv} : Le premier nom vernaculaire associe a l'espece, si disponible.
}
Si aucun ID ne peut etre determine, \code{id} sera \code{NA}. Si aucun nom vernaculaire n'est trouve,
\code{nv} sera \code{NA}.
}
\description{
Cette fonction permet de recuperer l'identifiant unique (ID) INPN ainsi que le nom vernaculaire
associe a une espece donnee en interrogeant l'API REST de l'INPN
}
\details{
La fonction effectue une recherche approximative (\emph{fuzzy matching}) sur le nom scientifique de l'espece
pour recuperer les informations depuis l'API TaxRef. Les sous-especes, varietes et formes, ainsi que les noms
illegitimes, sont exclus de la recherche.
}
\note{
Cette fonction necessite les packages \code{httr} et \code{jsonlite}. Si ces derniers ne sont pas installes,
la fonction s'arrete et invite l'utilisateur a les installer.
}
\examples{
\dontrun{
# Exemple d'utilisation
resultat <- match_inpn("Helianthemum ledifolium")
print(resultat$id)  # Affiche l'ID INPN
print(resultat$nv)  # Affiche le nom vernaculaire
}

}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}
}
