% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{vulnerabilite}
\alias{vulnerabilite}
\title{Calculer la vulnerabilite d'une espece}
\usage{
vulnerabilite(lr_nat, lr_reg = NA)
}
\arguments{
\item{lr_nat}{Statut dans la liste rouge nationale (\code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"DD"}, \code{"NA"}, \code{"NE"}).
Ce parametre est obligatoire.}

\item{lr_reg}{Statut dans la liste rouge regionale (\code{"LC"}, \code{"NT"}, \code{"VU"}, \code{"EN"}, \code{"CR"}, \code{"DD"}, \code{"NA"}, \code{"NE"}).
Par defaut, il est defini a \code{NA}.}
}
\value{
Un entier representant le niveau de vulnerabilite, avec une echelle de 1 (faible) a 5 (tres elevee).
Si \code{lr_reg} est \code{NA} ou egal a \code{"RE"} (\emph{region eteinte}), la vulnerabilite est calculee uniquement
en fonction du statut national. Sinon, une matrice de correspondance est utilisee pour combiner les statuts nationaux
et regionaux.
}
\description{
Cette fonction evalue le niveau de vulnerabilite d'une espece en fonction de son statut dans
les listes rouges nationales et regionales.
}
\details{
La fonction applique une logique en deux etapes :
\itemize{
\item Si le statut regional (\code{lr_reg}) est absent (\code{NA}) ou "RE", seule la vulnerabilite nationale est utilisee.
\item Sinon, une matrice predefinie est utilisee pour combiner les deux statuts et retourner un niveau de vulnerabilite.
}

Les codes des listes rouges sont interpretes comme suit :
\itemize{
\item \code{"LC"} : Preoccupation mineure.
\item \code{"NT"} : Quasi menace.
\item \code{"VU"} : Vulnerable.
\item \code{"EN"} : En danger.
\item \code{"CR"} : En danger critique.
\item \code{"DD"} : Donnees insuffisantes.
\item \code{"NA"} : Non applicable.
\item \code{"NE"} : Non evalue.
}
}
\examples{
\dontrun{
# Calcul de la vulnerabilite avec seulement le statut national
vul1 <- vulnerabilite("VU")
print(vul1)  # Retourne 3

# Calcul de la vulnerabilite avec les statuts national et regional
vul2 <- vulnerabilite("EN", "VU")
print(vul2)  # Retourne 4
}

}
