% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{calculer_ratios}
\alias{calculer_ratios}
\title{Calcul des ratios de representativite}
\usage{
calculer_ratios(id, departement)
}
\arguments{
\item{id}{Un identifiant numerique correspondant l'identifiant TaxRef de l'espece}

\item{departement}{Une chaine de caracteres specifiant le nom du departement pour lequel les
ratios doivent etre calcules.}
}
\value{
Une liste nommee contenant trois ratios :
\itemize{
\item \code{dep/reg} : Le ratio entre le nombre d'occurrences dans le departement et le nombre
d'occurrences dans la region.
\item \code{dep/nat} : Le ratio entre le nombre d'occurrences dans le departement et le nombre
d'occurrences au niveau national.
\item \code{reg/nat} : Le ratio entre le nombre d'occurrences dans la region et le nombre
d'occurrences au niveau national.
}
Si l'un des niveaux (departement, region ou national) ne contient aucune occurrence, le ratio
correspondant est defini comme \code{NA}.
}
\description{
Cette fonction calcule les ratios de representativite pour un taxon donne aux niveaux departemental,
regional et national en utilisant les donnees fournies par l'API REST OpenObs.
}
\details{
La fonction interroge l'API REST d'OpenObs pour recuperer le nombre d'occurrences d'un taxon specifique
aux niveaux departemental, regional et national. Ces valeurs sont ensuite utilisees pour calculer
les ratios de representativite.
}
\note{
Cette fonction necessite les packages \code{httr} et \code{jsonlite}. Si ces derniers ne sont pas installes,
la fonction s'arrete et invite l'utilisateur a les installer.
}
\examples{
\dontrun{
# Exemple d'utilisation
id = definir_id_inpn("Helianthemum ledifolium")

result <- calculer_ratios(id$id, departement = "aude")
print(result)
}

}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}
}
