% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{extraction_INPN}
\alias{extraction_INPN}
\title{Cette fonction permet de recuperer plusieurs informations d'interets de l'INPN sur une
espece donnee en interrogeant les API REST de l'INPN et de OpenObs}
\usage{
extraction_INPN(espece, departement)
}
\arguments{
\item{espece}{Une chaine de caracteres correspondant au nom scientifique de l'espece.}

\item{departement}{Une chaine de caracteres specifiant le nom du departement pour lequel les
statuts doivent etre recuperes.}
}
\value{
Un data.frame d'une ligne contenant :
\itemize{
\item \code{taxref} : l'identifiant TaxRef de l'espece.
\item \code{espece} : nom scientifique donne en parametre.
\item \code{nom.vernaculaires} : Le premier nom vernaculaire associe a l'espece, si disponible.
\item \code{dep.reg} : Le ratio entre le nombre d'occurrences dans le departement et le nombre
d'occurrences dans la region.
\item \code{dep.nat} : Le ratio entre le nombre d'occurrences dans le departement et le nombre
d'occurrences au niveau national.
\item \code{reg.nat} : Le ratio entre le nombre d'occurrences dans la region et le nombre
d'occurrences au niveau national.
\item \code{Les statuts ...} : Les informations sur le type de statut (protection, liste rouge, reglementation, ZNIEFF)...
\item \code{enjeux} : score d'enjeu de conservation d'une espece

}
}
\description{
Cette fonction permet de recuperer plusieurs informations d'interets de l'INPN sur une
espece donnee en interrogeant les API REST de l'INPN et de OpenObs
}
\details{
\itemize{
\item Pour les colonnes \code{taxref} et \code{nom.vernaculaire} voir \code{\link{definir_id_inpn}}.
\item Pour \code{dep.reg}, \code{dep.nat} et \code{reg.nat} voir \code{\link{calculer_ratios}}.
\item Pour \code{Les Statuts ...} voir \code{\link{recuperer_statuts}}.
\item Pour \code{enjeux} voir \code{\link{enjeu}}
}
}
\note{
Cette fonction necessite les packages \code{httr} et \code{jsonlite}. Si ces derniers ne sont pas installes,
la fonction s'arrete et invite l'utilisateur a les installer.
}
\examples{
\dontrun{
# Exemple d'utilisation
espece <- "Helianthemum ledifolium"
departement <- "aude"

res <- extraction_IPNI(espece, departement)
}
}
