% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{recuperer_statuts}
\alias{recuperer_statuts}
\title{Recuperation des statuts INPN d'un taxon}
\usage{
recuperer_statuts(id, departement)
}
\arguments{
\item{id}{Un identifiant numerique correspondant l'identifiant TaxRef de l'espece}

\item{departement}{Une chaine de caracteres specifiant le nom du departement pour lequel les
statuts doivent etre recuperes.}
}
\value{
Un tableau (\code{data.frame}) contenant les statuts filtres pour le taxon, avec les colonnes suivantes :
\itemize{
\item Les informations sur le type de statut (protection, liste rouge, reglementation, ZNIEFF).
\item Les localisations pertinentes (France metropolitaine, departement, ancienne et nouvelle region).
\item Une colonne supplementaire \code{col_names} combinant le type de statut et la localisation pour faciliter l'interpretation.
}
Les lignes non pertinentes sont exclues du resultat.
}
\description{
Cette fonction recupere les statuts INPN associes a un taxon donne pour un departement specifique,
ainsi que pour les anciennes et nouvelles regions correspondantes.
}
\details{
La fonction interroge l'API REST de l'INPN pour obtenir les statuts d'un taxon. Ces statuts sont filtres pour inclure
uniquement les localisations et categories d'interet (liste rouge, protection, ZNIEFF, reglementation).
}
\note{
Cette fonction necessite les packages \code{httr} et \code{jsonlite}. Si ces derniers ne sont pas installes,
la fonction s'arrete et invite l'utilisateur a les installer.
}
\examples{
\dontrun{
# Exemple d'utilisation
id = definir_id_inpn("Helianthemum ledifolium")

result <- recuperer_statuts(id$id, departement = "aude")
print(result)
}

}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}
}
